refactor: centralize data load/save
