+++ worker/my-worker/src/data.ts
@@ -1,0 +1,110 @@
+import type { Env } from './env';
+import { type Data, encryptField, decryptField } from './crypto';
+
+/**
+ * Load all bot data from the database.
+ */
+export async function loadData(env: Env): Promise<Data> {
+  const data: Data = { products: {}, pending: [], pending_add: [], pending_edit: [], languages: {} };
+
+  const prodRes = await env.DB.prepare('SELECT * FROM products').all();
+  for (const row of prodRes.results as any[]) {
+    const buyers = row.buyers ? JSON.parse(row.buyers) : [];
+    data.products[row.id] = {
+      price: row.price,
+      username: await decryptField(row.username, env.AES_KEY || env.FERNET_KEY),
+      password: await decryptField(row.password, env.AES_KEY || env.FERNET_KEY),
+      secret: await decryptField(row.secret, env.AES_KEY || env.FERNET_KEY),
+      buyers,
+    };
+    if (row.name) data.products[row.id].name = row.name;
+  }
+
+  const pendRes = await env.DB.prepare('SELECT user_id, product_id FROM pending').all();
+  data.pending = (pendRes.results as any[]).map((r) => ({
+    user_id: r.user_id,
+    product_id: r.product_id,
+  }));
+
+  const addRes = await env.DB.prepare('SELECT user_id, step, data FROM pending_add').all();
+  data.pending_add = (addRes.results as any[]).map((r) => ({
+    user_id: r.user_id,
+    step: r.step,
+    data: r.data ? JSON.parse(r.data) : {},
+  }));
+
+  const editRes = await env.DB.prepare('SELECT user_id, product_id, field FROM pending_edit').all();
+  data.pending_edit = (editRes.results as any[]).map((r) => ({
+    user_id: r.user_id,
+    product_id: r.product_id,
+    field: r.field,
+  }));
+
+  const langRes = await env.DB.prepare('SELECT user_id, lang FROM languages').all();
+  for (const row of langRes.results as any[]) {
+    data.languages[String(row.user_id)] = row.lang;
+  }
+
+  return data;
+}
+
+/**
+ * Persist the bot data back to the database.
+ */
+export async function saveData(env: Env, data: Data): Promise<void> {
+  const statements = [
+    env.DB.prepare('DELETE FROM products'),
+    env.DB.prepare('DELETE FROM pending'),
+    env.DB.prepare('DELETE FROM languages'),
+    env.DB.prepare('DELETE FROM pending_add'),
+    env.DB.prepare('DELETE FROM pending_edit'),
+  ];
+
+  for (const [id, product] of Object.entries(data.products)) {
+    const key = env.AES_KEY || env.FERNET_KEY;
+    const encUser = await encryptField(product.username, key);
+    const encPass = await encryptField(product.password, key);
+    const encSecret = await encryptField(product.secret, key);
+    const buyers = JSON.stringify(product.buyers || []);
+    statements.push(
+      env.DB.prepare(
+        'INSERT INTO products (id, price, username, password, secret, name, buyers) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)',
+      ).bind(id, product.price, encUser, encPass, encSecret, product.name ?? null, buyers),
+    );
+  }
+
+  for (const pending of data.pending) {
+    statements.push(
+      env.DB.prepare('INSERT INTO pending (user_id, product_id) VALUES (?1, ?2)').bind(pending.user_id, pending.product_id),
+    );
+  }
+
+  for (const add of data.pending_add) {
+    statements.push(
+      env.DB.prepare('INSERT INTO pending_add (user_id, step, data) VALUES (?1, ?2, ?3)').bind(
+        add.user_id,
+        add.step,
+        JSON.stringify(add.data),
+      ),
+    );
+  }
+
+  for (const edit of data.pending_edit) {
+    statements.push(
+      env.DB.prepare('INSERT INTO pending_edit (user_id, product_id, field) VALUES (?1, ?2, ?3)').bind(
+        edit.user_id,
+        edit.product_id,
+        edit.field,
+      ),
+    );
+  }
+
+  for (const [uid, lang] of Object.entries(data.languages)) {
+    statements.push(
+      env.DB.prepare('INSERT INTO languages (user_id, lang) VALUES (?1, ?2)').bind(Number(uid), lang),
+    );
+  }
+
+  await env.DB.batch(statements);
+}
+
